[WIP] Split rgb_matrix_indicators
