+++ quantum/rgb_matrix/rgb_matrix.h
@@ -119,6 +119,8 @@ enum rgb_matrix_effects {
     RGB_MATRIX_EFFECT_MAX
 };
 
+RGB rgb_matrix_hsv_to_rgb(HSV hsv);
+
 void eeconfig_update_rgb_matrix_default(void);
 void eeconfig_update_rgb_matrix(void);
 
@@ -131,16 +133,6 @@ void rgb_matrix_set_color_all(uint8_t red, uint8_t green, uint8_t blue);
 void process_rgb_matrix(uint8_t row, uint8_t col, bool pressed);
 
 void rgb_matrix_task(void);
-
-// This runs after another backlight effect and replaces
-// colors already set
-void rgb_matrix_indicators(void);
-bool rgb_matrix_indicators_kb(void);
-bool rgb_matrix_indicators_user(void);
-
-void rgb_matrix_indicators_advanced(effect_params_t *params);
-bool rgb_matrix_indicators_advanced_kb(uint8_t led_min, uint8_t led_max);
-bool rgb_matrix_indicators_advanced_user(uint8_t led_min, uint8_t led_max);
 
 void rgb_matrix_init(void);
 
@@ -270,10 +262,3 @@ extern last_hit_t g_last_hit_tracker;
 #ifdef RGB_MATRIX_FRAMEBUFFER_EFFECTS
 extern uint8_t g_rgb_frame_buffer[MATRIX_ROWS][MATRIX_COLS];
 #endif
-
-typedef struct {
-    int led_index;
-    led_t led_state;
-    HSV color;
-    bool override_brightness_limit;
-} t_rgb_indicator_matcher;
